+++ pkg/app/app.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"strconv"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
 	applicationv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
@@ -19,7 +18,7 @@ type Config struct {
 	AppName             string
 	AppNamespace        string
 	AppVersion          string
-	ConfigMajorVersion  int
+	ConfigVersion       string
 	DisableForceUpgrade bool
 	Name                string
 	UserConfigMapName   string
@@ -32,8 +31,8 @@ type Config struct {
 func NewCR(c Config) *applicationv1alpha1.App {
 	annotations := map[string]string{}
 	{
-		if c.ConfigMajorVersion > 0 {
-			annotations[annotation.ConfigMajorVersion] = strconv.Itoa(c.ConfigMajorVersion)
+		if c.ConfigVersion != "" {
+			annotations[annotation.ConfigVersion] = c.ConfigVersion
 		}
 		if !c.DisableForceUpgrade {
 			annotations["chart-operator.giantswarm.io/force-helm-upgrade"] = "true"